Add files via upload
This program imports a csv file and uses pandas module to calculate mean, median and standard deviation of 2 classes of students This is my 2nd attempt at using Github. Here, the program imports a csv file from a local computer. Then, it utilizes pandas module to analyse mean, median and standard deviation on grades of 2 classes ie Spring 2016 and Fall 2016.